requirements.txt
requests==2.19.1
python-dateutil==2.7.3
validate_email
rx==1.6.1